# reservas-apiv1/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Reservas API",
        "description": "API para sistema de reservas con pagos PayPal y autenticación Sanctum",
        "contact": {
            "email": "soporte@tusistema.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Servidor Local"
        }
    ],
    "paths": {
...
        "/paypal/create-order": {
            "post": {
                "tags": [
                    "Pagos"
                ],
                "summary": "Crear orden de pago en PayPal",
                "operationId": "createPayPalOrder",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "consultant_id",
                                    "date",
                                    "hours"
                                ],
                                "properties": {
                                    "consultant_id": {
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "date": {
                                        "type": "string",
                                        "example": "2024-10-23"
                                    },
                                    "hours": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "example": "10:00"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Orden creada correctamente"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/paypal/webhook": {
            "post": {
                "tags": [
                    "Pagos"
                ],
                "summary": "Webhook PayPal para confirmar pago",
                "description": "Endpoint llamado por PayPal cuando el pago se completa",
                "operationId": "19826948639c8df4f04e23de54395741",
                "responses": {
                    "200": {
                        "description": "Pago confirmado y reservas creadas correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Pago confirmado y reservas creadas"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error de validación o pago no completado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Pago no completado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Error en webhook"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Mensaje de excepción"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Ingrese el token en formato: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}